The controllers are defined, used a middleware and updated event.js script
