+++ app.js
@@ -6,6 +6,7 @@ const json = require('json');
 const authroutes = require('./routes/authroutes');
 const eventroutes = require('./routes/events');
 const adminroutes = require('./routes/admin');
+const preResponseMiddleware = require('./middleware/preResponseMiddleware');
 const PORT = process.env.PORT || 3000;
 const db = require('./db/connect');
 const method = require('method-override');
@@ -19,6 +20,7 @@ app.use(method('_method'));
 app.use(authroutes);
 app.use(eventroutes);
 app.use(adminroutes);
+app.use(preResponseMiddleware);
 app.listen(PORT, () => {
     console.log(`Server is running on port ${PORT}`);
 });
