+++ db/connect.js
@@ -1,4 +1,3 @@
-const express = require('express');
 const mongoose = require('mongoose');
 mongoose.connect('mongodb://127.0.0.1:27017/Event-Manager');
 if(mongoose.connection.readyState){
