+++ middleware/preResponseMiddleware.js
@@ -1,0 +1,11 @@
+// preresponse handler
+
+const preResponseMiddleware = (req, res, next) => {
+    res.handleServerError = (err) => {
+        console.error(err);
+        res.status(500).send('Internal Server Error');
+    };
+    next();
+}
+
+module.exports = preResponseMiddleware;
