+++ views/home.ejs
@@ -1,72 +1,174 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Home Page</title>
+
+    <style>
+        body {
+            font-family: 'Arial', sans-serif;
+            background-color: #f4f4f4;
+            margin: 0;
+            padding: 20px;
+        }
+
+        .navbar {
+            background-color: #3498db;
+            overflow: hidden;
+            border-radius: 8px;
+            padding: 10px;
+            margin-bottom: 20px;
+        }
+
+        .navbar a {
+            float: left;
+            display: block;
+            color: #fff;
+            text-align: center;
+            padding: 14px 16px;
+            text-decoration: none;
+        }
+
+        .navbar a:hover,
+        .dropdown:hover .dropbtn {
+            background-color: #1b5a7a;
+        }
+
+        .dropdown {
+            float: left;
+            overflow: hidden;
+        }
+
+        .dropdown .dropbtn {
+            font-size: 16px;
+            border: none;
+            outline: none;
+            color: #fff;
+            padding: 14px 16px;
+            background-color: inherit;
+            font-family: inherit;
+            margin: 0;
+        }
+
+        .dropdown-content {
+            display: none;
+            position: absolute;
+            background-color: #f9f9f9;
+            box-shadow: 0 8px 16px 0 rgba(0, 0, 0, 0.2);
+            z-index: 1;
+        }
+
+        .dropdown-content button {
+            color: #333;
+            padding: 12px 16px;
+            text-decoration: none;
+            display: block;
+            text-align: center;
+            background-color: #3498db;
+            border: none;
+            border-radius: 4px;
+            margin: 5px;
+            cursor: pointer;
+        }
+
+        .dropdown-content button:hover {
+            background-color: #1b5a7a;
+        }
+
+        .dropdown:hover .dropdown-content {
+            display: block;
+        }
+
+        p {
+            color: #333;
+            text-align: center;
+        }
+    </style>
 </head>
+
 <body>
     <% if (user) { %>
-        <p>Welcome, <%= user.username %>!</p>
-        <button type="button" id="logout">Logout</button> | <button type="button" id="profile">Profile</button>
-        <button type="button" id="events">Events</button>
-        <%if (user.role=='admin'){%>
-        <button type="button" id="dashboard">Analytics</button>
-        <%}%>
-    <% } else { %>
-        <% if (message) { %>
-            <p><%= message %></p>
-        <% } %>
-        <p>Welcome to the home page</p>
-        <button type="button" id="signUp">Sign Up</button> | <button type="button" id="signIn">Login</button>
-    <% } %>
+        <div class="navbar">
+            <p>Welcome, <%= user.username %>!</p>
+            <div class="dropdown">
+                <button class="dropbtn">Menu</button>
+                <div class="dropdown-content">
+                    <button id="logout">Logout</button>
+                    <button id="profile">Profile</button>
+                    <button id="events">Events</button>
+                    <% if (user.role=='admin' ) { %>
+                        <button id="dashboard">Analytics</button>
+                        <% } %>
+                </div>
+            </div>
+        </div>
+        <% } else { %>
+            <% if (message) { %>
+                <p>
+                    <%= message %>
+                </p>
+                <% } %>
+                    <p>Welcome to the home page</p>
+                    <div class="navbar">
+                        <div class="dropdown">
+                            <button class="dropbtn">Menu</button>
+                            <div class="dropdown-content">
+                                <button id="signUp">Sign Up</button>
+                                <button id="signIn">Login</button>
+                            </div>
+                        </div>
+                    </div>
+                    <% } %>
 
-    <script>
-        //button click event for sign up and sign in
-    const signUpButton = document.getElementById('signUp');
-    const signInButton = document.getElementById('signIn');
-    const logoutButton = document.getElementById('logout');
-    const profileButton = document.getElementById('profile');
-    const events = document.getElementById('events');
-    const Analytics= document.getElementById('dashboard');
+                        <p>Explore the Event Management System for seamless event planning and organization.</p>
 
-    if (signUpButton) {
-        signUpButton.addEventListener('click', () => {
-            window.location.href = "/register";
-        });
-    }
+                        <script>
+                            //button click event for sign up, sign in, and navigation
+                            const signUpButton = document.getElementById('signUp');
+                            const signInButton = document.getElementById('signIn');
+                            const logoutButton = document.getElementById('logout');
+                            const profileButton = document.getElementById('profile');
+                            const events = document.getElementById('events');
+                            const Analytics = document.getElementById('dashboard');
 
-    if (signInButton) {
-        signInButton.addEventListener('click', () => {
-            window.location.href = "/login";
-        });
-    }
+                            if (signUpButton) {
+                                signUpButton.addEventListener('click', () => {
+                                    window.location.href = "/register";
+                                });
+                            }
 
-    if (logoutButton) {
-        logoutButton.addEventListener('click', () => {
-            window.location.href = "/logout";
-        });
-    }
+                            if (signInButton) {
+                                signInButton.addEventListener('click', () => {
+                                    window.location.href = "/login";
+                                });
+                            }
 
-    if (profileButton) {
-        profileButton.addEventListener('click', () => {
-            window.location.href = "/profile";
-        });}
+                            if (logoutButton) {
+                                logoutButton.addEventListener('click', () => {
+                                    window.location.href = "/logout";
+                                });
+                            }
 
-        if(events){
-            events.addEventListener('click',() =>{
-                window.location.href="/events"
-       
-            });
+                            if (profileButton) {
+                                profileButton.addEventListener('click', () => {
+                                    window.location.href = "/profile";
+                                });
+                            }
 
-     }
-     if(Analytics){
-        Analytics.addEventListener('click',()=>{
-            window.location.href="/analytics"
-        });
-     }
-    
+                            if (events) {
+                                events.addEventListener('click', () => {
+                                    window.location.href = "/events";
+                                });
+                            }
 
-    </script>
+                            if (Analytics) {
+                                Analytics.addEventListener('click', () => {
+                                    window.location.href = "/analytics";
+                                });
+                            }
+                        </script>
 </body>
+
 </html>
