+++ views/login.ejs
@@ -5,7 +5,75 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Login</title>
-    <!-- Add your CSS stylesheets or include a stylesheet framework like Bootstrap -->
+    <style>
+        body {
+            font-family: 'Arial', sans-serif;
+            background-color: #f4f4f4;
+            margin: 0;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            height: 100vh;
+        }
+    
+        h2 {
+            text-align: center;
+            color: #333;
+        }
+    
+        form {
+            background-color: #fff;
+            padding: 20px;
+            border-radius: 8px;
+            box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+            width: 300px;
+        }
+    
+        label {
+            display: block;
+            margin-bottom: 8px;
+            color: #333;
+        }
+    
+        input {
+            width: 100%;
+            padding: 8px;
+            margin-bottom: 16px;
+            box-sizing: border-box;
+            border: 1px solid #ccc;
+            border-radius: 4px;
+        }
+    
+        button {
+            background-color: #3498db;
+            color: #fff;
+            padding: 10px;
+            border: none;
+            border-radius: 4px;
+            cursor: pointer;
+            width: 100%;
+        }
+    
+        button:hover {
+            background-color: #1b5a7a;
+        }
+    
+        p {
+            text-align: center;
+            margin-top: 16px;
+            color: #666;
+        }
+    
+        a {
+            color: #3498db;
+            text-decoration: none;
+        }
+    
+        a:hover {
+            text-decoration: underline;
+        }
+    </style>
+    
 </head>
 <body>
     <h2>
@@ -22,10 +90,8 @@
 
         <label for="password">Password:</label>
         <input type="password" id="password" name="password" required>
-
+        <p>Don't have an account? <a href="/register">Register</a></p>
         <button type="submit">Login</button>
     </form>
-
-    <p>Don't have an account? <a href="/register">Register</a></p>
 </body>
 </html>
